(feat) Database ajax calls to store and retrieve past searches
