+++ client/src/components/SearchBox.jsx
@@ -4,6 +4,37 @@ class SearchBox extends React.Component {
     this.state = {value: ''};
   }
 
+  saveSearch(){
+    //For testing purposes. Want to ensure Heroku database works
+
+    //save search
+    $.ajax({
+      url: '/api/searches',
+      type: 'POST',
+      data: { query: this.state.value },
+      success: function(data) {
+        console.log('success')
+      }.bind(this),
+      error: function() {
+        console.log('failure')
+      }
+    });
+  }
+
+  listSearch(){
+    //grab all past searches
+    $.ajax({
+      url: '/api/searches',
+      type: 'GET',
+      success: function(data) {
+        console.log('success' + JSON.stringify(data))
+      }.bind(this),
+      error: function() {
+        console.log('failure')
+      }
+    });
+  }
+
   searchAPI() {
     this.props.searchAPI({query: this.state.value}, (recipes) => {
       this.props.callback(recipes);
@@ -16,7 +47,7 @@ class SearchBox extends React.Component {
         <input className="searchbox" type="text" onChange={(event) => {
           this.setState({value: event.target.value});
         }} value={this.state.value}/>
-        <button className="submitButton" onClick={() => {this.searchAPI()}}>Search</button>
+        <button className="submitButton" onClick={() => {this.searchAPI(); this.saveSearch(); this.listSearch();}}>Search</button>
       </div>
     )
   }
