+++ package.json
@@ -32,7 +32,8 @@
     "nodemon": "^1.2.1",
     "react": "^0.14.7",
     "react-dom": "^0.14.7",
-    "request": "^2.36.0"
+    "request": "^2.36.0",
+    "mongoose": "~4.4"
   },
   "devDependencies": {
     "eslint": "^3.1.1",
